extract path into a new component
